[http] Import/export request tracking ID.
This commit implements request ID tracking through HTTP.

HTTPClient is a new struct that wraps *http.Client.  It's Do
method extracts the current request ID from the context of the
passed *http.Request, and add a special HTTP header with the
request ID value.

The default HTTP header is "X-Cybozu-Request-ID".  If there
is "REQUEST_ID_HEADER" environment variable, its value is used
as the header instead.

HTTPClient also records logs.  If logged in JSON format,
the log can be unmarshalled in RequestLog.

HTTPServer is modified to import such a request ID from the
header if exists.  The request ID is then added to the request
context value.