[exec] record command execution results.
This commit implements LogCmd that is a wrapper for *exec.Cmd.

Similar to HTTPCient, LogCmd records command execution results
and optionally includes the tracking request ID if given context
has a request ID value.

CommandContext is a function similar to exec.CommandContext, but
creates and resturns *LogCmd instead of *exec.Cmd.