Add support for HTTP/2 server.
The ResponseWriter in net/http package implements a different set
of methods for HTTP/1 and HTTP/2.  Specifically, ResponseWriter
for HTTP/2 does not implement http.Hijacker and io.ReaderFrom,
but it implements http.Pusher interface.

In addition to well.StdResponseWriter that is HTTP/1, this
adds well.StdResponseWriter2 for HTTP/2.

well.HTTPServer now returns ResponseWriter that implements
a set of methods defined in well.StdResponseWriter2 for HTTP/2
requests.